refactor switch to own file
