+++ parser/src/cfg/mod.rs
@@ -57,6 +57,9 @@ pub use defcfg::*;
 mod deftemplate;
 pub use deftemplate::*;
 
+mod switch;
+pub use switch::*;
+
 use crate::custom_action::*;
 use crate::keys::*;
 use crate::layers::*;
@@ -2980,142 +2983,6 @@ fn parse_sequence_start(ac_params: &[SExpr], s: &ParsedState) -> Result<&'static
     Ok(s.a.sref(Action::Custom(s.a.sref(
         s.a.sref_slice(CustomAction::SequenceLeader(timeout, input_mode)),
     ))))
-}
-
-fn parse_switch(ac_params: &[SExpr], s: &ParsedState) -> Result<&'static KanataAction> {
-    const ERR_STR: &str =
-        "switch expects triples of params: <key match> <action> <break|fallthrough>";
-
-    let mut cases = vec![];
-
-    let mut params = ac_params.iter();
-    loop {
-        let Some(key_match) = params.next() else {
-            break;
-        };
-        let Some(action) = params.next() else {
-            bail!("{ERR_STR}\nMissing <action> and <break|fallthrough> for the final triple");
-        };
-        let Some(break_or_fallthrough_expr) = params.next() else {
-            bail!("{ERR_STR}\nMissing <break|fallthrough> for the final triple");
-        };
-
-        let Some(key_match) = key_match.list(s.vars()) else {
-            bail_expr!(key_match, "{ERR_STR}\n<key match> must be a list")
-        };
-        let mut ops = vec![];
-        for op in key_match.iter() {
-            parse_switch_case_bool(1, op, &mut ops, s)?;
-        }
-
-        let action = parse_action(action, s)?;
-
-        let Some(break_or_fallthrough) = break_or_fallthrough_expr.atom(s.vars()) else {
-            bail_expr!(
-                break_or_fallthrough_expr,
-                "{ERR_STR}\nthis must be one of: break, fallthrough"
-            );
-        };
-        let break_or_fallthrough = match break_or_fallthrough {
-            "break" => BreakOrFallthrough::Break,
-            "fallthrough" => BreakOrFallthrough::Fallthrough,
-            _ => bail_expr!(
-                break_or_fallthrough_expr,
-                "{ERR_STR}\nthis must be one of: break, fallthrough"
-            ),
-        };
-        cases.push((s.a.sref_vec(ops), action, break_or_fallthrough));
-    }
-    Ok(s.a.sref(Action::Switch(s.a.sref(Switch {
-        cases: s.a.sref_vec(cases),
-    }))))
-}
-
-fn parse_switch_case_bool(
-    depth: u8,
-    op_expr: &SExpr,
-    ops: &mut Vec<OpCode>,
-    s: &ParsedState,
-) -> Result<()> {
-    if ops.len() > MAX_OPCODE_LEN as usize {
-        bail_expr!(
-            op_expr,
-            "maximum key match size of {MAX_OPCODE_LEN} items is exceeded"
-        );
-    }
-    if usize::from(depth) > MAX_BOOL_EXPR_DEPTH {
-        bail_expr!(
-            op_expr,
-            "maximum key match expression depth {MAX_BOOL_EXPR_DEPTH} is exceeded"
-        );
-    }
-    if let Some(a) = op_expr.atom(s.vars()) {
-        let osc = str_to_oscode(a).ok_or_else(|| anyhow_expr!(op_expr, "invalid key name"))?;
-        ops.push(OpCode::new_key(osc.into()));
-        Ok(())
-    } else {
-        let l = op_expr
-            .list(s.vars())
-            .expect("must be a list, checked atom");
-        if l.is_empty() {
-            bail_expr!(op_expr, "key match cannot contain empty lists inside");
-        }
-        #[derive(PartialEq)]
-        enum AllowedListOps {
-            Or,
-            And,
-            Not,
-            KeyHistory,
-        }
-        let op = l[0]
-            .atom(s.vars())
-            .and_then(|s| match s {
-                "or" => Some(AllowedListOps::Or),
-                "and" => Some(AllowedListOps::And),
-                "not" => Some(AllowedListOps::Not),
-                "key-history" => Some(AllowedListOps::KeyHistory),
-                _ => None,
-            })
-            .ok_or_else(|| {
-                anyhow_expr!(
-                    op_expr,
-                    "lists inside key match must begin with one of: or | and | not | key-history"
-                )
-            })?;
-        match op {
-            AllowedListOps::KeyHistory => {
-                if l.len() != 3 {
-                    bail_expr!(
-                        op_expr,
-                        "key-history must have 2 parameters: key, key-recency"
-                    );
-                }
-                let osc = l[1]
-                    .atom(s.vars())
-                    .and_then(str_to_oscode)
-                    .ok_or_else(|| anyhow_expr!(op_expr, "invalid key name"))?;
-                let key_recency = parse_u8_with_range(&l[2], s, "key-recency", 1, 8)? - 1;
-                ops.push(OpCode::new_key_history(osc.into(), key_recency));
-                Ok(())
-            }
-            AllowedListOps::Or | AllowedListOps::And | AllowedListOps::Not => {
-                let op = match op {
-                    AllowedListOps::Or => BooleanOperator::Or,
-                    AllowedListOps::And => BooleanOperator::And,
-                    AllowedListOps::Not => BooleanOperator::Not,
-                    _ => unreachable!(),
-                };
-                // insert a placeholder for now, don't know the end index yet.
-                let placeholder_index = ops.len() as u16;
-                ops.push(OpCode::new_bool(op, placeholder_index));
-                for op in l.iter().skip(1) {
-                    parse_switch_case_bool(depth + 1, op, ops, s)?;
-                }
-                ops[placeholder_index as usize] = OpCode::new_bool(op, ops.len() as u16);
-                Ok(())
-            }
-        }
-    }
 }
 
 fn parse_on_idle_fakekey(ac_params: &[SExpr], s: &ParsedState) -> Result<&'static KanataAction> {
